Use a simple avro-encoded protocol
Rather than using some specific avro messages, just avro-encode the
schemas as an array as the first message down the pipe. Then all
messages past that are an index into that array to identify the schema
followed by an array of encoded schema values. This will certainly need
to grow a bit, but it's good enough to get data down the wire to start
with.