Attempt to shutdown the store in an orderly fashion when the server is shutdown
First the series parcelers stop creating new splits when united writers are closed. Then all channels to clients are closed, which closes all the writers. Finally the store waits a minute for any active splits to finish up, terminating them if they don't finish in that window. This shuts things down as quickly as possible after the shutdown command is received.

When the server next starts up, any writers that weren't split at the end will be split on the startup check. It'd be nice to move that splitting onto a background thread to make startup faster, but that's for another day.